+++ RELEASE_NOTES.md
@@ -61,10 +61,10 @@ _________________
 
 1. Install server:
 ```bash
-docker run -d -p 8050:8050 -p 8080:8080 backupsystem/server:latest
+docker run -d -p 8000:8000 backupsystem/server:latest
 ```
 
-2. Access UI: http://localhost:8080
+2. Access API/Swagger: http://localhost:8000/swagger
 
 3. Install agent:
 ```bash